refactor: input을 모두 inputView에서 처리
